Basic install and service
Installs from a tarball that is fetched from GitHub. This is far from ideal.
We're just using it for testing at this stage. You probably don't want to
use this for anything serious as indicated by the README.

I'm installing to a sub-directory so that we stand some change of this
working if we need to change the version number.